+++ R/bigrquery.R
@@ -14,6 +14,8 @@
 #' @importFrom Rcpp sourceCpp
 #' @importFrom tibble tibble
 #' @importFrom bit64 integer64
+#' @importFrom rlang %||%
+#' @importFrom glue glue glue_data glue_collapse
 #' @keywords internal
 #' @useDynLib bigrquery, .registration = TRUE
 "_PACKAGE"