+++ R/old-dataset.R
@@ -9,12 +9,18 @@ NULL
 #' @export
 #' @rdname dataset-dep
 get_dataset <- function(project, dataset) {
+
+  .Deprecated("bq_dataset_meta", package = "bigrquery")
+
   bq_get(bq_path(project, dataset))
 }
 
 #' @export
 #' @rdname dataset-dep
 exists_dataset <- function(project, dataset) {
+
+  .Deprecated("bq_dataset_exists", package = "bigrquery")
+
   tryCatch(
     !is.null(get_dataset(project = project, dataset = dataset)),
     bigrquery_notFound = function(e) FALSE
@@ -24,6 +30,9 @@ exists_dataset <- function(project, dataset) {
 #' @export
 #' @rdname dataset-dep
 delete_dataset <- function(project, dataset, deleteContents = FALSE) {
+
+  .Deprecated("bq_dataset_delete", package = "bigrquery")
+
   assert_that(is.string(project), is.string(dataset))
 
   url <- sprintf("projects/%s/datasets/%s", project, dataset)
@@ -33,6 +42,9 @@ delete_dataset <- function(project, dataset, deleteContents = FALSE) {
 #' @export
 #' @rdname dataset-dep
 insert_dataset <- function(project, dataset, ...) {
+
+  .Deprecated("bq_dataset_create", package = "bigrquery")
+
   assert_that(is.string(project), is.string(dataset))
 
   url <- bq_path(project, "")
@@ -49,6 +61,9 @@ insert_dataset <- function(project, dataset, ...) {
 #' @export
 #' @rdname dataset-dep
 update_dataset <- function(project, dataset, ...) {
+
+  .Deprecated("bq_dataset_update", package = "bigrquery")
+
   url <- bq_path(project, dataset)
   body <- list(
     datasetReference = list(
@@ -63,6 +78,9 @@ update_dataset <- function(project, dataset, ...) {
 #' @export
 #' @rdname dataset-dep
 list_tables <- function(project, dataset, page_size = 50, max_pages = Inf) {
+
+  .Deprecated("bq_dataset_tables", package = "bigrquery")
+
   data <- bq_get_paginated(
     bq_path(project, dataset, ""),
     page_size = page_size,