+++ R/old-table.R
@@ -8,6 +8,7 @@ NULL
 #' @rdname table-dep
 #' @export
 insert_table <- function(project, dataset, table, ...) {
+  .Deprecated("bq_table_upload", package = "bigrquery")
   x <- bq_table(project, dataset, table)
   bq_table_create(x, ...)
 }
@@ -15,6 +16,7 @@ insert_table <- function(project, dataset, table, ...) {
 #' @rdname table-dep
 #' @export
 get_table <- function(project, dataset, table) {
+  .Deprecated("bq_table_meta", package = "bigrquery")
   x <- bq_table(project, dataset, table)
   bq_table_meta(x)
 }
@@ -22,6 +24,7 @@ get_table <- function(project, dataset, table) {
 #' @rdname table-dep
 #' @export
 exists_table <- function(project, dataset, table) {
+  .Deprecated("bq_table_exists", package = "bigrquery")
   x <- bq_table(project, dataset, table)
   bq_table_exists(x)
 }
@@ -29,7 +32,7 @@ exists_table <- function(project, dataset, table) {
 #' @rdname table-dep
 #' @export
 delete_table <- function(project, dataset, table) {
+  .Deprecated("bq_table_delete", package = "bigrquery")
   x <- bq_table(project, dataset, table)
   bq_table_delete(x)
 }
-