+++ R/utils.R
@@ -5,8 +5,6 @@ as_df <- function(x) {
   x
 }
 
-"%||%" <- function(x, y) if (is.null(x)) y else x
-
 bq_quiet <- function(x) {
   if (is.na(x)) {
     !interactive()
@@ -32,7 +30,19 @@ bq_progress <- function(..., quiet = NA) {
   }
 }
 
+bq_check_namespace <- function(pkg, bq_type) {
+  if (requireNamespace(pkg, quietly = TRUE)) {
+    return()
+  }
+
+  rlang::abort(glue::glue(
+    "Package '{pkg}' must be installed to load BigQuery field with type '{bq_type}'"
+  ))
+}
+
 isFALSE <- function(x) identical(x, FALSE)
+
+is_string <- function(x) length(x) == 1L && is.character(x)
 
 cat_line <- function(...) {
   cat(paste0(..., "\n", collapse = ""))
@@ -64,6 +74,6 @@ show_json <- function(x) {
 
 #' @export
 print.bq_bytes <- function(x, ...) {
-  cat_line(prettyunits::pretty_bytes(x))
+  cat_line(prettyunits::pretty_bytes(unclass(x)))
 }
 # nocov end