+++ tests/testthat.R
@@ -1,4 +1,4 @@
-library("testthat")
-library("bigrquery")
+library(testthat)
+library(bigrquery)
 
 test_check("bigrquery")