+++ tests/testthat/helper-auth.R
@@ -1,0 +1,4 @@
+if (gargle:::secret_can_decrypt("bigrquery")) {
+  json <- gargle:::secret_read("bigrquery", "bigrquery-testing.json")
+  bq_auth(path = rawToChar(json))
+}