+++ tests/testthat/test-bq-parse.R
@@ -1,5 +1,3 @@
-context("test-bq-parse.R")
-
 # Individual values -------------------------------------------------------
 
 test_that("can parse atomic vectors", {
@@ -125,8 +123,8 @@ test_that("can parse arrays of structs", {
 # Complete files ----------------------------------------------------------
 
 replay_query <- function(name, sql) {
-  schema_path <- test_path(glue::glue("parse-schema-{name}.json"))
-  values_path <- test_path(glue::glue("parse-values-{name}.json"))
+  schema_path <- test_path(glue("parse-schema-{name}.json"))
+  values_path <- test_path(glue("parse-values-{name}.json"))
 
   if (!file.exists(schema_path)) {
     tbl <- bq_project_query(bq_test_project(), sql)
@@ -175,3 +173,26 @@ test_that("can parse empty arrays", {
   df <- bq_table_download(tb)
   expect_equal(df$x, list(tibble::tibble(a = integer(length = 0), b = character())))
 })
+
+test_that("can parse geography", {
+  skip_if_not_installed("wk")
+
+  wkt <- wk::wkt("POINT (30 10)")
+  expect_identical(bq_parse_single(as.character(wkt), "geography"), wkt)
+  expect_identical(bq_parse_single(NA_character_, "geography"), wk::wkt(NA_character_))
+})
+
+test_that("can parse bytes", {
+  bytes <- blob::blob(
+    as.raw(c(0x01, 0x01, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff,
+             0xff, 0xff, 0x3d, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x24,
+             0x40))
+  )
+
+  expect_identical(bq_parse_single(bytes[[1]], "bytes"), bytes)
+  expect_identical(
+    bq_parse_single(NA_character_, "bytes"),
+    blob::blob(NULL)
+  )
+})
+